+++ internal/usecase/user_usecase/register_or_get.go
@@ -6,10 +6,10 @@ import (
 	"ping_prog/internal/domain"
 )
 
-func (uc *UseCase) RegisterOrGet(ctx context.Context, username string, chatID int64) (*domain.User, error) {
+func (uc *UseCase) RegisterOrGet(ctx context.Context, username string, chatID int) (*domain.User, error) {
 	user := domain.NewUser(username, chatID)
 
-	_ = uc.userRepo.Create(ctx, user) // игнорируем ошибку, если уже существует
+	_ = uc.userRepo.Create(ctx, user) // ignore error if already exists
 
 	return uc.userRepo.GetByChatID(ctx, chatID)
 }
